DD-66 ♻ [Refactor] Singup 폴더 내 CommonLayout을 FormLayout으로 수정
