--- src/features/Signup/index.ts
@@ -1,5 +1,0 @@
-export {
-  CommonLayout,
-  CenterLayout, Label,
-  BottomWrapper,
-} from './layout/CommonLayout';
