+++ src/pages/signup/SignUpProfilePage.tsx
@@ -1,8 +1,12 @@
 import ProgressBar from '@/components/Bar/ProgressBar';
 import Button from '@/components/Button/Button';
 import { PROFILE_COLORS } from '@/constants/colors';
-import { BottomWrapper, CommonLayout, Label } from '@/features/Signup';
-import { ProgressBarWrapper } from '@/features/Signup/layout/CommonLayout';
+import {
+  ProgressBarWrapper,
+  BottomWrapper,
+  CommonLayout,
+  Label,
+} from '@/layouts/FormLayoutStyles';
 import { useSignUpStore } from '@/store/useSignupStore';
 import styled from '@emotion/styled';
 import React from 'react';
